requirements.txt
numpy~=1.19.5
matplotlib~=3.3.4
Pillow~=8.0.1
torchvision~=0.8.2
torch~=1.7.1
PyYAML~=5.4.1
pandas~=1.2.3
tensorboard~=2.4.0
tqdm~=4.51.0
sklearn~=0.0
scikit-learn~=0.24.1
joblib~=1.0.0